Add files via upload
# Kickstarting with Excel

## Overview of Project
The purpuse of this challneg was to receieve a large set of data and firstly, review how it is set up and what information is has collected. From there, the challenge was to add to the data set to make it easier to understand and start making calculations to help draw conclusions that would eventualy be communicated in a graph.

### Purpose

## Analysis and Challenges
From our excersice I was about to separate the parent categories and subcategories and convert the orginial launch date number to actualy dates and eventually to years. From there I was able to create a pivot chart showing the outcomes of all of the compaigns based on launch date. I filtered by the theater category and created a line graph showing how well the campaigns performed by month. 

### Analysis of Outcomes Based on Launch Date
My analysis on the outcomes base on launch date would be that ingeneral theater campaigns are not very successful and are therefore a risky move. 
### Analysis of Outcomes Based on Goals

### Challenges and Difficulties Encountered

## Results

- What are two conclusions you can draw about the Outcomes based on Launch Date?
The best time to start a campaign is in January and the worst time to start a campaign is in June.
- What can you conclude about the Outcomes based on Goals?

- What are some limitations of this dataset?
The data set does not show the current economic status of each year. If would be helpful to see if the economy was strong of weak during the time of sucessful and failed campaigns. The data also does not show where in each country the campign what failed or succeeded took place or the marketing aspects of the campaign. 
- What are some other possible tables and/or graphs that we could create?
We could show where categories of campaigned performed the best and worst.